moved tests to tests/regex_options.rs
