+++ src/lib.rs
@@ -1685,7 +1685,7 @@ mod tests {
     use alloc::{format, vec};
 
     use crate::parse::make_literal;
-    use crate::{Expr, Regex, RegexBuilder};
+    use crate::{Expr, Regex};
 
     //use detect_possible_backref;
 
@@ -1771,49 +1771,6 @@ mod tests {
         assert_eq!(to_str(repeat(0, usize::MAX, false)), "a*?");
         assert_eq!(to_str(repeat(1, usize::MAX, true)), "a+");
         assert_eq!(to_str(repeat(1, usize::MAX, false)), "a+?");
-    }
-
-    #[test]
-    fn check_casing_option() {
-        let builder = RegexBuilder::new(r"TEST foo")
-            .case_insensitive(false)
-            .build();
-
-        match builder {
-            Ok(regex) => {
-                assert!(regex.is_match(r"TEST foo").unwrap_or_default());
-                assert!(!regex.is_match(r"test foo").unwrap_or_default());
-            }
-            _ => panic!("builder should be able to compile with casing options"),
-        }
-    }
-
-    #[test]
-    fn check_override_casing_option() {
-        let builder = RegexBuilder::new(r"FOO(?i:bar)quux")
-            .case_insensitive(false)
-            .build();
-
-        match builder {
-            Ok(regex) => {
-                assert!(!regex.is_match("FoObarQuUx").unwrap_or_default());
-                assert!(!regex.is_match("fooBARquux").unwrap_or_default());
-                assert!(regex.is_match("FOObarquux").unwrap_or_default());
-            }
-            _ => panic!("builder should be able to compile with casing options"),
-        }
-    }
-
-    #[test]
-    fn check_casing_insensitive_option() {
-        let builder = RegexBuilder::new(r"TEST FOO")
-            .case_insensitive(true)
-            .build();
-
-        match builder {
-            Ok(regex) => assert!(regex.is_match(r"test foo").unwrap_or_default()),
-            _ => panic!("builder should be able to compile with casing options"),
-        }
     }
 
     #[test]
